minor code structure change
