+++ lib/components/player_details.dart
@@ -1,5 +1,5 @@
 import 'package:flutter/material.dart';
-import 'package:pocket_eleven/player.dart';
+import 'package:pocket_eleven/controller/player.dart';
 import 'package:pocket_eleven/design/colors.dart';
 
 class PlayerDetailsDialog extends StatelessWidget {
