+++ lib/pages/club/club_youth_page.dart
@@ -1,18 +1,18 @@
 import 'package:flutter/material.dart';
 import 'package:pocket_eleven/design/colors.dart';
-import 'package:pocket_eleven/user_manager.dart';
+import 'package:pocket_eleven/controller/user_manager.dart';
 import 'package:unicons/unicons.dart';
 
-class ClubTrainingPage extends StatefulWidget {
+class ClubYouthPage extends StatefulWidget {
   final VoidCallback onCurrencyChange;
 
-  const ClubTrainingPage({super.key, required this.onCurrencyChange});
+  const ClubYouthPage({super.key, required this.onCurrencyChange});
 
   @override
-  State<ClubTrainingPage> createState() => _ClubTrainingPageState();
+  State<ClubYouthPage> createState() => _ClubYouthPageState();
 }
 
-class _ClubTrainingPageState extends State<ClubTrainingPage> {
+class _ClubYouthPageState extends State<ClubYouthPage> {
   late Image _clubStadiumImage;
   int level = 1;
   int upgradeCost = 100000;
@@ -20,9 +20,9 @@ class _ClubTrainingPageState extends State<ClubTrainingPage> {
   @override
   void initState() {
     super.initState();
-    _clubStadiumImage = Image.asset('assets/background/club_training.png');
-    level = UserManager.trainingLevel;
-    upgradeCost = UserManager.trainingUpgradeCost;
+    _clubStadiumImage = Image.asset('assets/background/club_youth.png');
+    level = UserManager.youthLevel;
+    upgradeCost = UserManager.youthUpgradeCost;
   }
 
   void increaseLevel() {
@@ -30,16 +30,16 @@ class _ClubTrainingPageState extends State<ClubTrainingPage> {
       setState(() {
         level++;
         UserManager.money -= upgradeCost;
-        UserManager.trainingLevel = level;
-        UserManager.trainingUpgradeCost =
+        UserManager.youthLevel = level;
+        UserManager.youthUpgradeCost =
             ((upgradeCost * 1.8) / 10000).round() * 10000;
-        upgradeCost = UserManager.trainingUpgradeCost;
+        upgradeCost = UserManager.youthUpgradeCost;
       });
 
       widget.onCurrencyChange();
 
-      UserManager().saveTrainingLevel();
-      UserManager().saveTrainingUpgradeCost();
+      UserManager().saveYouthLevel();
+      UserManager().saveYouthUpgradeCost();
     }
   }
 
@@ -100,7 +100,7 @@ class _ClubTrainingPageState extends State<ClubTrainingPage> {
               child: Column(
                 crossAxisAlignment: CrossAxisAlignment.start,
                 children: [
-                  _buildTrainingInfo(),
+                  _buildYouthInfo(),
                   SizedBox(height: screenHeight * 0.04),
                   const Text(
                     'Description',
@@ -114,9 +114,9 @@ class _ClubTrainingPageState extends State<ClubTrainingPage> {
                   const Expanded(
                     child: SingleChildScrollView(
                       child: Text(
-                        'Our state-of-the-art training facilities are the heart of our club infrastructure. Equipped with the latest technologies, '
-                        'they cater to the needs of both professional athletes and young talents. Here, under the supervision of our experts, '
-                        'players hone their skills, preparing for the most significant challenges on the field.',
+                        "Our youth academies are where future football stars develop their skills under the guidance of experienced coaches. "
+                        "We provide an inspiring environment for learning and nurturing a passion for soccer, "
+                        "shaping not just athletic abilities but also teamwork and determination.",
                         style: TextStyle(
                           fontSize: 16.0,
                           color: AppColors.textEnabledColor,
@@ -149,7 +149,7 @@ class _ClubTrainingPageState extends State<ClubTrainingPage> {
     );
   }
 
-  Widget _buildTrainingInfo() {
+  Widget _buildYouthInfo() {
     return Row(
       mainAxisAlignment: MainAxisAlignment.spaceBetween,
       children: [
@@ -158,7 +158,7 @@ class _ClubTrainingPageState extends State<ClubTrainingPage> {
             crossAxisAlignment: CrossAxisAlignment.start,
             children: [
               const Text(
-                'Training Facilities',
+                'Youth Academy',
                 style: TextStyle(
                   fontSize: 24.0,
                   fontWeight: FontWeight.bold,
